+++ src/test/java/com/strangegrotto/wealthdraft/assets/definition/AssetDefinitionsDeserializerTest.java
@@ -1,0 +1,6 @@
+package com.strangegrotto.wealthdraft.assets.definition;
+
+import org.junit.Test;
+
+public class AssetDefinitionsDeserializerTest {
+}